Review comments
PR #133
