+++ boa/util/abi.py
@@ -1,5 +1,5 @@
 # wrapper module around whatever encoder we are using
-from typing import TYPE_CHECKING, Any
+from typing import Annotated, Any
 
 from eth.codecs.abi import nodes
 from eth.codecs.abi.decoder import Decoder
@@ -7,29 +7,56 @@ from eth.codecs.abi.encoder import Encoder
 from eth.codecs.abi.exceptions import ABIError
 from eth.codecs.abi.nodes import ABITypeNode
 from eth.codecs.abi.parser import Parser
+from eth_typing import Address as PYEVM_Address
+from eth_utils import to_canonical_address, to_checksum_address
 
-if TYPE_CHECKING:
-    from boa.environment import Address
+from boa.util.lrudict import lrudict
 
 _parsers: dict[str, ABITypeNode] = {}
 
 
-class AbiEncoder(Encoder):
+# XXX: inherit from bytes directly so that we can pass it to py-evm?
+# inherit from `str` so that ABI encoder / decoder can work without failing
+class Address(str):  # (PYEVM_Address):
+    # converting between checksum and canonical addresses is a hotspot;
+    # this class contains both and caches recently seen conversions
+    __slots__ = ("canonical_address",)
+    _cache = lrudict(1024)
+
+    canonical_address: Annotated[PYEVM_Address, "canonical address"]
+
+    def __new__(cls, address):
+        if isinstance(address, Address):
+            return address
+
+        try:
+            return cls._cache[address]
+        except KeyError:
+            pass
+
+        checksum_address = to_checksum_address(address)
+        self = super().__new__(cls, checksum_address)
+        self.canonical_address = to_canonical_address(address)
+        cls._cache[address] = self
+        return self
+
+    def __repr__(self):
+        checksum_addr = super().__repr__()
+        return f"_Address({checksum_addr})"
+
+
+class ABIEncoder(Encoder):
     @classmethod
     def visit_AddressNode(cls, node: nodes.AddressNode, value) -> bytes:
         value = getattr(value, "address", value)
         return super().visit_AddressNode(node, value)
 
 
-class AbiDecoder(Decoder):
+class ABIDecoder(Decoder):
     @classmethod
     def visit_AddressNode(
         cls, node: nodes.AddressNode, value: bytes, checksum: bool = True, **kwargs: Any
     ) -> "Address":
-        from boa.environment import (  # Maybe Address should be in a different file?
-            Address,
-        )
-
         ret = super().visit_AddressNode(node, value)
         return Address(ret)
 
@@ -43,11 +70,11 @@ def _get_parser(schema: str):
 
 
 def abi_encode(schema: str, data: Any) -> bytes:
-    return AbiEncoder.encode(_get_parser(schema), data)
+    return ABIEncoder.encode(_get_parser(schema), data)
 
 
 def abi_decode(schema: str, data: bytes) -> Any:
-    return AbiDecoder.decode(_get_parser(schema), data)
+    return ABIDecoder.decode(_get_parser(schema), data)
 
 
 def is_abi_encodable(abi_type: str, data: Any) -> bool:
